Fix issues with Long in schema (#1191)

+++ examples/server/ktor-server/src/main/kotlin/com/expediagroup/graphql/examples/server/ktor/schema/UniversityQueryService.kt
@@ -26,4 +26,4 @@ class UniversityQueryService : Query {
         University.search(params.ids)
 }
 
-data class UniversitySearchParameters(val ids: List<Long>)
+data class UniversitySearchParameters(val ids: List<Int>)
