+++ examples/server/ktor-server/src/main/kotlin/com/expediagroup/graphql/examples/server/ktor/schema/dataloaders/BookDataLoader.kt
@@ -22,9 +22,9 @@ import kotlinx.coroutines.runBlocking
 import org.dataloader.DataLoader
 import java.util.concurrent.CompletableFuture
 
-val BookDataLoader = object : KotlinDataLoader<List<Long>, List<Book>> {
+val BookDataLoader = object : KotlinDataLoader<List<Int>, List<Book>> {
     override val dataLoaderName = "BATCH_BOOK_LOADER"
-    override fun getDataLoader() = DataLoader<List<Long>, List<Book>> { ids ->
+    override fun getDataLoader() = DataLoader<List<Int>, List<Book>> { ids ->
         CompletableFuture.supplyAsync {
             val allBooks = runBlocking { Book.search(ids.flatten()).toMutableList() }
             // produce lists of results from returned books
